+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # HEAD
+
+## Bugfixes
+- Correct typo `ray_interception` -> `ray_intersection`
 - Wrong operator in `ray_interception(pl::Plane, ray::Ray)` [#12](https://github.com/stevelonny/jujutracer/issues/7)
 
 # Version 0.2.0